requirements.txt
Django==1.6.5
MySQL-python==1.2.4
Pillow==2.4.0
South==0.8.2
Werkzeug==0.9.4
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.32
celery==3.0.23
django-crispy-forms==1.4.0
django-extensions==1.3.7
django-recaptcha==0.0.6
easy-thumbnails==2.0.1
ipdb==0.8
ipython==2.1.0
kombu==2.5.14
python-dateutil==1.5
python-magic==0.4.3
readline==6.2.4.1
six==1.4.1
wsgiref==0.1.2
django-pagination-bootstrap==1.0.5
-e git+git@github.com:laysakura/rainbow_logging_handler.git@2425d00f7dc562e23881b7bab2abe9d0128bf46d#egg=rainbow_logging_handler
requests==2.3.0
redis==2.9.1
recaptcha-client==1.0.6
django-nose==1.2
django-redis
uwsgi==2.0.3
django-registration==1.0
flake8
djrill==1.0.0
huey==0.4.2
mock==1.0.1
unicode-slugify==0.1.1
django-autoslug==1.7.2